% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{diagnostic-generic}
\alias{diagnostic}
\alias{diagnostic-generic}
\title{null assumption diagnostic plots}
\usage{
diagnostic(obj, ...)
}
\description{
null assumption diagnostic plots
}
\keyword{internal}

