\name{diagnostics}
\alias{diagnostics}
\title{diagnostic plots for hsigclust}
\usage{
diagnostics(x, plot = FALSE, ...)
}
\description{
defining new diagnostics() method for hsigclust class also:
this needs to be cleaned up and separated into multiple
plotting functions e.g. diagnostic plots and pval plots
should be separate.
}
\author{
Patrick Kimes
}

