% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{eigval_dat}
\alias{eigval_dat}
\title{eigval_dat
return \code{eigval_dat} slot}
\usage{
eigval_dat(obj)
}
\arguments{
\item{obj}{object with slot \code{eigval_dat}}
}
\description{
eigval_dat
return \code{eigval_dat} slot
}
\keyword{internal}

