% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{eigval_sim}
\alias{eigval_sim}
\title{eigval_sim
return \code{eigval_sim} slot}
\usage{
eigval_sim(obj)
}
\arguments{
\item{obj}{object with slot \code{eigval_sim}}
}
\description{
eigval_sim
return \code{eigval_sim} slot
}
\keyword{internal}

