% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fwer_cutoff,list,numeric-method}
\alias{fwer_cutoff,list,numeric-method}
\title{compute cutoffs based on hierachical clustering output from idx_hc}
\usage{
\S4method{fwer_cutoff}{list,numeric}(obj, alpha)
}
\description{
compute cutoffs based on hierachical clustering output from idx_hc
}
\keyword{internal}

