% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fwer_cutoff-matrix}
\alias{fwer_cutoff-matrix}
\alias{fwer_cutoff.matrix}
\title{get FWER from idx_hc attribute of shc object}
\usage{
\method{fwer_cutoff}{shc}(obj, alpha, ...)
}
\arguments{
\item{obj}{\code{shc} object}

\item{alpha}{numeric value specifying level}

\item{...}{other parameters to be used by the function}
}
\description{
get FWER from idx_hc attribute of shc object
}
\author{
Patrick Kimes
}
\keyword{internal}

