% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shc-fwer_control.R
\name{fwer_cutoff-shc}
\alias{fwer_cutoff-shc}
\alias{fwer_cutoff.shc}
\title{get FWER cutoffs for shc oject}
\usage{
\method{fwer_cutoff}{shc}(obj, alpha, ...)
}
\arguments{
\item{obj}{\code{shc} object}

\item{alpha}{numeric value specifying level}

\item{...}{other parameters to be used by the function}
}
\description{
get FWER cutoffs for shc oject
}
\author{
Patrick Kimes
}

