% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fwer_cutoff}
\alias{fwer_cutoff}
\title{fwer_cutoff
return Family-Wise Error Rate (FWER) cutoffs}
\usage{
fwer_cutoff(obj, alpha)
}
\arguments{
\item{obj}{currently only for objects of class \code{shc-class}}

\item{alpha}{a numeric value between 0 and 1}
}
\description{
fwer_cutoff
return Family-Wise Error Rate (FWER) cutoffs
}
\keyword{internal}

