\name{hsigclust-controlFWER}
\alias{hsigclust-controlFWER}
\title{HSigClust FWER controlling procedure}
\usage{
controlFWER(hsc, alpha = 0.05, ...)
}
\description{
Calculate the sequence of p-value cutoffs for FWER control
as described in Meinshausen (2008)
}
\details{
some details
}
\author{
Patrick Kimes
}
\references{
Meinshausen, N. (2008). Hierarchical testing of variable
importance. Biometrika, 95(2), 265-278.
}

