\name{hsigclust-diagnostics}
\alias{hsigclust-diagnostics}
\title{plot sigclust diagnostic plots for hsigclust object}
\usage{
diagnostics(x, plot = FALSE, ...)
}
\description{
generates various sigclust diagnostic plots defined in the
sigclust package for evaluating the Gaussian null
covariance assumptions of the SigClust hypothesis test
}
\details{
some details
}
\author{
Patrick Kimes
}

