\docType{methods}
\name{hsigclust-getters}
\alias{asimcindex}
\alias{clusterList}
\alias{hc}
\alias{hsigclust-getters}
\alias{icovest}
\alias{meigval}
\alias{mpval}
\alias{mpvalnorm}
\alias{msimeigval}
\alias{nsim}
\alias{vsimbackvar}
\alias{xmcindex}
\title{getters for the hsigclust class}
\usage{
raw.data(x)

meigval(x)

msimeigval(x)

vsimbackvar(x)

icovest(x)

nsim(x)

asimcindex(x)

mpval(x)

mpvalnorm(x)

xmcindex(x)

clusterList(x)

hc(x)
}
\value{
the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis

the original data used in the analysis
}
\description{
Defining getters (no setters) for the hsigclust class.
}
\author{
Patrick Kimes
}

