\docType{package}
\name{HSigClust-package}
\alias{HSigClust}
\alias{HSigClust-package}
\title{significance tests for clustering in HDLSS settings}
\description{
\code{\link{HSCtest}} computes \eqn{p}-values for
statistical significance of clustering in hierarchical
clustering.
}
\details{
some details
}
\author{
Patrick Kimes \email{pkimes@live.unc.edu}
}
\keyword{package}

