\name{hsigclust-plot}
\alias{hsigclust-plot}
\title{plot hsigclust object}
\usage{
.plot.hsigclust(hsigclust, arg = "all", ...)
}
\description{
visualize results of HSigClust analysis as an annotated
dendrogram object using the ggdendro and ggplot2 packages
}
\details{
some details
}
\author{
Patrick Kimes
}

