\docType{methods}
\name{hsigclust-print}
\alias{hsigclust-print}
\title{print hsigclust object}
\usage{
\S4method{print}{hsigclust}(x, arg = "all", ...)
}
\arguments{
  \item{x}{an object used to select a method.}

  \item{...}{further arguments passed to or from other
  methods.}
}
\description{
quick output to screen when hsigclust object is evaluated
}
\details{
some details
}
\author{
Patrick Kimes
}

