\docType{methods}
\name{hsigclust-show}
\alias{hsigclust-show}
\title{show hsigclust object}
\usage{
\S4method{show}{hsigclust}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
quick output to screen when hsigclust object is evaluated.
I'm not exactly sure how this differs from the \code{print}
method. Currently under investigation, but defining both
for safety.
}
\details{
some details
}
\author{
Patrick Kimes
}

