\docType{methods}
\name{hsigclust-summary}
\alias{hsigclust-summary}
\title{summary for hsigclust object}
\usage{
\S4method{summary}{hsigclust}(object, ...)
}
\arguments{
  \item{object}{an object for which a summary is desired.}

  \item{...}{additional arguments affecting the summary
  produced.}
}
\description{
details on the results of the HSigClust testing procedure
}
\details{
some details
}
\author{
Patrick Kimes
}

