% Generated by roxygen2 (4.0.2): do not edit by hand
\name{null_eigval}
\alias{null_eigval}
\title{Eigenvalue estimation for null Gaussian based testing procedures}
\usage{
null_eigval(x, n, p, icovest = 1, bkgd_pca = TRUE)
}
\arguments{
\item{x}{a matrix of size n by p containing the original data}

\item{n}{an integer number of samples}

\item{p}{an integer number of features/covariates}

\item{icovest}{an integer between 1 and 3 corresponding to the covariance
estimation procedure to use. See details for more
information on the possible estimation procedures (default = 1)}

\item{bkgd_pca}{a logical value specifying whether to use scaled PCA scores
       or raw data to estimate the background noise (default = TRUE)}
}
\value{
The function returns a list of estimated parameters for the null Gaussian
distribution used in significance of clustering testing. The list includes:
\itemize{
    \item \code{eigval_dat}: eigenvalues for sample covariance matrix
    \item \code{backvar}: background noise, sigma_b^2
    \item \code{eigval_sim}: eigenvalues to be used for simulation
}
}
\description{
function to compute the eigenvalues of the null Gaussian distribution for
significance of clustering testing procedures which rely on a null Gaussian
factor model assumption
}
\details{
The following possible options are given for null covariance estimation
\enumerate{
    \item soft thresholding: recommended approach described in Huang et al. 2014
    \item sample: uses sample covariance matrix, equivalent to soft and hard options
          when n > p, but when p > n, will produce conservative results, i.e. less
          significant p-values
    \item hard thresholding: approach described in Liu et al. 2008, no longer
          recommended - retained for historical purposes
}
}
\author{
Patrick Kimes
}
\references{
\itemize{
    \item Huang, H., Liu, Y., Yuan, M., and Marron, J. S. (2014).
          Statistical Significance of Clustering using Soft Thresholding.
          Journal of Computational and Graphical Statistics, preprint.
    \item Liu, Y., Hayes, D. N., Nobel, A. B., and Marron, J. S. (2008).
          Statistical Significance of Clustering for High-Dimension, Low-Sample Size Data.
          Journal of the American Statistical Association, 103(483):1281-1293.
}
}

