% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shc-print.R
\name{print-shc}
\alias{print-shc}
\alias{print.shc}
\title{print description of shc object}
\usage{
\method{print}{shc}(x, ...)
}
\arguments{
\item{x}{a \code{shc} object}

\item{...}{other parameters to be used by the function}
}
\description{
Provides information about \code{shc}
}
\author{
Patrick Kimes
}
