\name{setValidity}
\alias{setValidity}
\title{Validity check for the hsigclust class}
\usage{
.validtype(object)
}
\description{
making sure the \code{hsigclust} object is correctly
constructed
}
\details{
some details
}
\author{
Patrick Kimes
}
\references{
http://www.bioconductor.org/help/course-materials/2010/AdvancedR/S4InBioconductor.pdf
(slide 12)
}

