% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{class}
\name{shc-class}
\alias{backvar,shc-method}
\alias{ci_dat,shc-method}
\alias{ci_sim,shc-method}
\alias{eigval_dat,shc-method}
\alias{eigval_sim,shc-method}
\alias{fwer_cutoff,shc,numeric-method}
\alias{hc_dat,shc-method}
\alias{idx_hc,shc-method}
\alias{in_args,shc-method}
\alias{in_mat,shc-method}
\alias{nd_type,shc-method}
\alias{p_emp,shc-method}
\alias{p_norm,shc-method}
\alias{shc-class}
\alias{show,shc-method}
\alias{summary,shc-method}
\title{significance for hierarchical clustering (shc) object}
\usage{
\S4method{in_mat}{shc}(obj)

\S4method{in_args}{shc}(obj)

\S4method{eigval_dat}{shc}(obj)

\S4method{eigval_sim}{shc}(obj)

\S4method{backvar}{shc}(obj)

\S4method{nd_type}{shc}(obj)

\S4method{ci_sim}{shc}(obj)

\S4method{ci_dat}{shc}(obj)

\S4method{p_emp}{shc}(obj)

\S4method{p_norm}{shc}(obj)

\S4method{idx_hc}{shc}(obj)

\S4method{hc_dat}{shc}(obj)

\S4method{fwer_cutoff}{shc,numeric}(obj, alpha)

\S4method{show}{shc}(object)

\S4method{summary}{shc}(object, ...)
}
\arguments{
\item{obj}{\code{shc} object}
}
\description{
S4 class encapsulating various elements of a SHC analysis.
}
\section{Methods (by generic)}{
\itemize{
\item \code{in_mat}: get \code{in_mat} slot

\item \code{in_args}: get \code{in_args} slot

\item \code{eigval_dat}: get \code{eigval_dat} slot

\item \code{eigval_sim}: get \code{eigval_sim} slot

\item \code{backvar}: get \code{backvar} slot

\item \code{nd_type}: get \code{nd_type} slot

\item \code{ci_sim}: get \code{ci_sim} slot

\item \code{ci_dat}: get \code{ci_dat} slot

\item \code{p_emp}: get \code{p_emp} slot

\item \code{p_norm}: get \code{p_norm} slot

\item \code{idx_hc}: get \code{idx_hc} slot

\item \code{hc_dat}: get \code{hc_dat} slot

\item \code{fwer_cutoff}: return FWER cutoffs based on \code{idx_hc} slot

\item \code{show}: show brief summary of \code{shc} object

\item \code{summary}: show brief summary of shc object
}}
\section{Slots}{

\describe{
\item{\code{in_mat}}{the original data matrix passed to the constructor}

\item{\code{in_args}}{a list of the original parameters passed to the constructor,
see \code{shc} for list of parameters}

\item{\code{eigval_dat}}{a matrix containing the sample eigenvalues for each subtree
tested along the dendrogram}

\item{\code{eigval_sim}}{a matrix containing the estimated eigenvalues used to
simulate null data at each subtree tested along the dendrogram}

\item{\code{backvar}}{a vector containing the estimated background variances used for
computing \code{eigval_sim}}

\item{\code{nd_type}}{a vector of length n-1 taking values in "\code{n_small}",
"\code{no_test}", "\code{tested}" specifying how each node along the
dendrogram was handled by the iterative testing procedure}

\item{\code{ci_dat}}{a matrix containing the cluster indices for the original
data matrix passed to the constructor}

\item{\code{ci_sim}}{a 3-dimensional array containing the simulated cluster
indices at each subtree tested along the dendrogram}

\item{\code{pval_emp}}{a matrix containing the emprical p-values computed at each
subtree tested along the dendrogram}

\item{\code{pval_norm}}{a matrix containing the Gaussian approximate p-values
computed at each subtree tested along the dendrogram}

\item{\code{idx_hc}}{a list of tuples containing the indices of clusters joined
at each step of the hierarchical clustering procedure}

\item{\code{hc_dat}}{a \code{hclust} object constructed from the original data matrix and
      arguments passed to the constructor}
}}
\author{
Patrick Kimes
}

