% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{shc-constructor}
\alias{shc,matrix-method}
\alias{shc-constructor}
\title{statistical Significance for Hierarchical Clustering (SHC)
algorithm}
\usage{
shc(x, ...)

\S4method{shc}{matrix}(x, metric, linkage, l = 2, alpha = 1, icovest = 1,
  bkgd_pca = TRUE, n_sim = 100, n_min = 10, rcpp = TRUE, ci = "2CI",
  ci_null = "hclust", ci_idx = 1, ci_emp = TRUE)
}
\arguments{
\item{x}{a dataset with n rows and p columns, with observations in rows}

\item{metric}{a string specifying the metric to be used in the hierarchical
clustering procedure. This must be a metric accepted by \code{dist},
e.g. "euclidean," or "cor."}

\item{linkage}{a string specifying the linkage to be used in the hierarchical
clustering procedure. This must be a linkage accepted by
\code{hclust}, e.g. "ward."}

\item{l}{an integer value specifying the power of the Minkowski distance, if
used (default = 2)}

\item{alpha}{a value between 0 and 1 specifying the desired level of the
test. If no FWER control is desired, simply set alpha to 1 (default = 1)}

\item{icovest}{a numeric value between 1 and 3 specifying the null covariance
estimation method to be used. See \code{\link{null_eigval}} for more
details (default = 1)}

\item{bkgd_pca}{a logical value whether to use principal component scores when
estimating background noise under the null (default = TRUE)}

\item{n_sim}{a numeric value specifying the number of simulations for Monte Carlo
testing (default = 100)}

\item{n_min}{an integer specifying the minimum number of observations needed
to calculate a p-value (default = 10)}

\item{rcpp}{a logical value whether to use the \code{Rclusterpp} package
(default = TRUE)}

\item{ci}{a string vector specifying the cluster indices to be used for
testing along the dendrogram. Currently, options include: "2CI",
"linkage". (default = "2CI")}

\item{ci_null}{a string vector specifying the clustering approach that
should be used at each node as the comparison. Currently, options
include: "2means", "hclust". Note, \code{ci_null} and \code{ci} must be of
equal length. (default = "hclust")}

\item{ci_idx}{a numeric value between 1 and \code{length(ci)}
specifiying which CI to use for the FWER stopping rule.
This only has an effect if \code{alpha} is specified to a non-default
value. (default = 1)}

\item{ci_emp}{a logical value specifying whether to use the empirical
       p-value from the CI based on \code{ci_idx} for the FWER stopping rule.
       As with \code{ci_idx} this only has an effect if \code{alpha} is
       specified to a non-default value. (default = TRUE)}

\item{verb}{a logical value specifying whether the method should print out
when testing completes along each node along the dendrogram
(default = FALSE)}
}
\value{
The function returns a \code{shc} object containing the
        resulting p-values. The print method call will output a dendrogram
        with the corresponding p-values placed at each merge.
}
\description{
implements the Monte Carlo simulation based
significance testing procedure described in Kimes et al. (pre-print).
Statistical significance is evaluated at each node along the tree
starting from the root using a Gaussian null hypothesis test.
A corresponding family-wise error rate (FWER) controlling procedure is
provided.
}
\details{
The function expands on the \code{sigclust} idea to the hierarchical
         setting by modifying the clustering procedure employed to compute
         the null distribution of cluster indices.

The code is written so that various cluster indices can be
easily introduced by writting new CI function and adding to
.initcluster(), .simcluster(). When possible, this function
makes use of a C++ implementation of hierarchical clustering
available through the \code{Rclusterpp.hclust} package for the
case of clustering by Pearson correlation (\code{dist="cor"}),
we make use of \code{WGCNA::cor} with the usual
code{stats::hclust}.

       The testing procedure will terminate when no branches
       meet the corresponding FWER control threshold. This procedure may
       substantially speed up the procedure by reducing the number of tests
       considered. NOTE: if p-values are desired for all branches, the FWER
       cutoffs may be provided a posteriori by a call to \code{FWERcutoffs()}.
       Additionally, they may be specified at \code{plot()} using the
       \code{alpha} and \code{FWER} parameters.
}
\examples{
hsc_cars <- shc(as.matrix(mtcars), metric="euclidean", linkage="single")
tail(p_norm(hsc_cars), 10)
}
\author{
Patrick Kimes
}
\references{
Kimes, P. K., Hayes, D. N., Liu Y., and Marron, J. S. (2014)
Statistical Significance for Hierarchical Clustering.
pre-print.
}
\seealso{
\code{\link{plot-shc}} \code{\link{diagnostic}}
}

