% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shcutree.R
\name{shcutree}
\alias{shcutree}
\title{Cut SHC Tree by Significance}
\usage{
shcutree(obj, alpha = 0.05, ci_idx = 1, ci_emp = FALSE)
}
\arguments{
\item{obj}{\code{shc} object}

\item{alpha}{a double between 0 and 1 specifying the FWER controlled significance
cutoff (default = 0.05)}

\item{ci_idx}{a numeric value between 1 and \code{length(ci)} 
specifiying which CI to use for the FWER stopping rule.
This only has an effect if \code{alpha} < 1. (default = 1)}

\item{ci_emp}{a logical value specifying whether to use the empirical
p-value from the CI based on \code{ci_idx} for the FWER stopping rule.
As with \code{ci_idx} this only has an effect if \code{alpha} < 1.
(default = FALSE)}
}
\value{
a vector of length n specifying cluster labels based on specified FWER
alpha threshold.
}
\description{
Returns the cluster labels based on cutting the tree at the specified
\code{alpha} level
}
\author{
Patrick Kimes
}
