% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sigclust2-constructor.R
\name{sigclust2}
\alias{sigclust2}
\alias{sigclust2-constructor}
\title{Statistical Significance of Clustering (sigclust)}
\usage{
sigclust2(x, n_sim = 100, n_start = 1, icovest = 1, bkgd_pca = TRUE,
  labels = NULL)
}
\arguments{
\item{x}{a dataset with n rows and p columns, with observations in rows}

\item{n_sim}{a numeric value specifying the number of simulations for Monte Carlo
testing (default = 100)}

\item{n_start}{a numeric value specifying the number of random starts to be used
for k-means clustering passed to \code{kmeans} (default = 1)}

\item{icovest}{an integer between 1 and 3 specifying the null covariance
estimation method to be used. See \code{\link{null_eigval}} for more
details (default = 1)}

\item{bkgd_pca}{a logical value whether to use principal component scores when
estimating background noise under the null (default = TRUE)}

\item{labels}{a n-vector of 1s and 2s specifying cluster labels for testing
       instead of using \code{kmeans} (default = NULL)}
}
\value{
The function returns a \code{sigclust2} S3-object containing the
resulting p-values. The \code{sigclust2} object has following attributes:
\itemize{
\item{\code{in_mat}}: {the original data matrix passed to the constructor}
}
}
\description{
Re-implementation of the Monte Carlo simulation based
significance testing procedure described in Liu et al. (2008)
}
\author{
Patrick Kimes
}
\references{
\itemize{
    \item Kimes, P. K., Hayes, D. N., Liu Y., and Marron, J. S. (2014)
          Statistical significance for hierarchical clustering.
          pre-print available.
}
}

