% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{summary-shc}
\alias{summary}
\alias{summary,shc-method}
\alias{summary-shc}
\title{provide summary of shc object}
\usage{
summary(object, ...)

\S4method{summary}{shc}(object, ...)
}
\arguments{
\item{object}{a \code{shc} object}
}
\description{
Provides information about \code{shc} object similar to call to \code{show}
}
\author{
Patrick Kimes
}

