% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summary.R
\name{summary-sigclust2}
\alias{summary-sigclust2}
\alias{summary.sigclust2}
\title{provide summary of sigclust2 object}
\usage{
\method{summary}{sigclust2}(object, ...)
}
\arguments{
\item{object}{a \code{sigclust2} object}

\item{...}{other parameters to be used by the function}
}
\description{
Provides information about \code{sigclust2} object similar to call to \code{show}
}
\author{
Patrick Kimes
}

