% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_hsc}
\alias{theme_hsc}
\title{hsigclust dendrogram theme}
\usage{
theme_hsc(base_size = 12)
}
\arguments{
\item{base_size}{a numeric value specifying base size of text}
}
\description{
\code{ggplot2} theme for plotting hsigclust output, currently just using
\code{theme_bw()}
}
\author{
Patrick Kimes
}

