% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diagnostic-shc}
\alias{diagnostic-shc}
\alias{diagnostic.shc}
\title{plot diagnostics for shc object}
\usage{
\method{diagnostic}{shc}(obj, K = nrow(obj$p_norm) - 1, fname = NULL,
  ci_idx = 1, ...)
}
\arguments{
\item{obj}{a \code{shc} object}

\item{K}{an integer value specifying the range of nodes for which to
create diagnostic plots (default = nrow(obj$p_norm)-1)}

\item{fname}{a character string specifying the name of the output file,
see details for more information on default behavior depending on
\code{length(K)} (default = NULL)}

\item{ci_idx}{a numeric value between 1 and \code{length(obj$in_args$ci)}
specifying which cluster index to use for creating plots (default = 1)}

\item{arg}{other parameters to be passed to function}
}
\value{
prints plots to \code{fname}.
}
\description{
Provides visualizations to check the null Gaussian assumption at a specified
range of nodes along the dendrogram
}
\details{
If \code{K} is a single value, the default behavior is to output the diagnostic
plot to the current device. This behavior is overriden if \code{fname} is specified
by the user. If \code{K} includes several values, than the diagnostic plots are
returned to \code{fname}.pdf. If \code{fname} is not specified, a default value,
"shc_diagnostic", is used.
}
\author{
Patrick Kimes
}

