% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shc-diagnostic.R
\name{diagnostic-shc}
\alias{diagnostic-shc}
\alias{diagnostic.shc}
\title{plot diagnostics for shc object}
\usage{
\method{diagnostic}{shc}(obj, K = 1, fname = NULL, ci_idx = 1,
  pty = "all", ...)
}
\arguments{
\item{obj}{a \code{shc} object}

\item{K}{an integer value or vector specifying the range of nodes for which to 
create diagnostic plots, with 1 corresponding to the root (default = 1)}

\item{fname}{a character string specifying the name of the output file,
see details for more information on default behavior depending on
\code{length(K)} (default = NULL)}

\item{ci_idx}{a numeric value between 1 and \code{length(obj$in_args$ci)}
specifying which cluster index to use for creating plots (default = 1)}

\item{pty}{a string specifying the plot type that should be created
see details for more information on different plot types available
(default = "all")}

\item{...}{other parameters passed to plots}
}
\value{
Prints plots to \code{paste0(fname, ".pdf")} if \code{fname} is specified or \code{length(K)>1}.
}
\description{
Provides visualizations to check the null Gaussian assumption at a specified
range of nodes along the dendrogram
}
\details{
If \code{K} is a single value, the default behavior is to output the diagnostic
plot to the current device. This behavior is overriden if \code{fname} is specified
by the user. If \code{length(K) > 1}, than the diagnostic plots are
printed to \code{paste0(fname, ".pdf")}. If \code{fname} is not specified,
"shc_diagnostic", is used by default.

The \code{pty} parameter accepts the following options:
\itemize{
\item{\code{"background"}}: {empirical distribution of marginal data used for background variance estimation}
\item{\code{"qq"}}: {qqplot for checking null Gaussian assumption}
\item{\code{"diag"}}: {screeplot for checking covariance factor model}
\item{\code{"pvalue"}}: {empirical distribution of simulated CIs}
\item{\code{"all"}}: {all of the above plots (default).}
}
}
\author{
Patrick Kimes
}

