% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{fwer_cutoff-generic}
\alias{fwer_cutoff}
\alias{fwer_cutoff-generic}
\title{return Family-Wise Error Rate (FWER) cutoffs}
\usage{
fwer_cutoff(obj, ...)
}
\description{
return Family-Wise Error Rate (FWER) cutoffs
}
\keyword{internal}

