\docType{methods}
\name{hsigclust-show}
\alias{hsigclust-show}
\title{show hsigclust object}
\usage{
\S4method{show}{hsigclust}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
Quick output to screen when hsigclust object is evaluated.
}
\details{
some details...
}
\author{
Patrick Kimes
}

