% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shc-constructor.R
\name{shc}
\alias{shc}
\alias{shc-constructor}
\title{statistical Significance for Hierarchical Clustering (SHC)}
\usage{
shc(x, metric = "euclidean", vecmet = NULL, matmet = NULL,
  linkage = "ward.D2", l = 2, alpha = 1, icovest = 1,
  bkgd_pca = FALSE, n_sim = 100, n_min = 10, rcpp = FALSE, ci = "2CI",
  null_alg = "hclust", ci_idx = 1, ci_emp = FALSE)
}
\arguments{
\item{x}{a dataset with n rows and p columns, with observations in rows and
features in columns.}

\item{metric}{a string specifying the metric to be used in the hierarchical 
clustering procedure. This must be a metric accepted by \code{dist}
or "cor" (to specify 1 - Pearson correlation). (default = "euclidean")}

\item{vecmet}{a function taking two vectors as input and returning a real-valued
number which specifies how dissimilarities should be computed between rows
of the data matrix (\code{x}). If non-NULL, will take precedence over
\code{metric}. Only one of \code{vecmet} or \code{matmet} can be specified.
(default = NULL)}

\item{matmet}{a function taking a matrix as input and returning an object of class
\code{dist} representing dissimilarities between rows of the data matrix
(\code{x}). If non-NULL, will take precedence over \code{metric}.
Only one of \code{vecmet} or \code{matmet} can be specified.
(default = NULL)}

\item{linkage}{a string specifying the linkage to be used in the hierarchical 
clustering procedure. This must be a linkage accepted by 
\code{Rclusterpp.hclust} if \code{rcpp=TRUE}, e.g. "ward",
or \code{stats::hclust} if \code{rcpp=FALSE}, e.g. "ward.D2".
(default = "ward.D2")}

\item{l}{an integer value specifying the power of the Minkowski distance, if 
used. (default = 2)}

\item{alpha}{a value between 0 and 1 specifying the desired level of the 
test. If no FWER control is desired, simply set alpha to 1. (default = 1)}

\item{icovest}{an integer (1, 2 or 3) specifying the null covariance
estimation method to be used. See \code{\link{null_eigval}} for more
details. (default = 1)}

\item{bkgd_pca}{a logical value specifying whether to use scaled PCA scores
over raw data to estimate the background noise. When FALSE, raw estimate
is used; when TRUE, minimum of PCA and raw estimates is used.
(default = FALSE)}

\item{n_sim}{a numeric value specifying the number of simulations at each node
for Monte Carlo testing. (default = 100)}

\item{n_min}{an integer specifying the minimum number of observations needed
to calculate a p-value. (default = 10)}

\item{rcpp}{a logical value whether to use the \code{Rclusterpp} package.
(default = FALSE)}

\item{ci}{a string vector specifying the cluster indices to be used for 
testing along the dendrogram. Currently, options include: "2CI", 
"linkage". (default = "2CI")}

\item{null_alg}{a string vector specifying the clustering algorithms that 
should be used to cluster the simulated null ditributions. Currently, options
include: "2means" and "hclust". While "hclust" typically provides greater power
for rotation invariant combinations of metric and linkage function, 
if a non-rotation invariant metric, e.g. Pearson correlation, is used it is
recommended that the "2means" option is specified. Note, \code{null_alg} and
\code{ci} must be of equal length. (default = "hclust")}

\item{ci_idx}{an integer between 1 and \code{length(ci)} 
specifiying which CI to use for the FWER stopping rule.
This only has an effect if \code{alpha} is specified to a non-default 
value. (default = 1)}

\item{ci_emp}{a logical value specifying whether to use the empirical
p-value from the CI based on \code{ci_idx} for the FWER stopping rule.
As with \code{ci_idx}, this only has an effect if \code{alpha} is
specified to a non-default value. (default = FALSE)}
}
\value{
The function returns a \code{shc} S3-object containing the 
resulting p-values. The \code{plot} method can be used to generate a dendrogram
with the corresponding p-values placed at each merge. The \code{shc}
object has following attributes:
\itemize{
\item{\code{in_mat}}: {the original data matrix passed to the constructor}
\item{\code{in_args}}: {a list of the original parameters passed to the constructor}
\item{\code{eigval_dat}}: {a matrix with each row containing the sample eigenvalues for a
    subtree tested along the dendrogram}
\item{\code{eigval_sim}}: {a matrix with each row containing the estimated eigenvalues used to
    simulate null data at a subtree tested along the dendrogram}
\item{\code{backvar}}: {a vector containing the estimated background variances used for
    computing \code{eigval_sim}}
\item{\code{nd_type}}: {a vector of length n-1 taking values in "\code{n_small}",
    "\code{no_test}", "\code{sig}", "\code{not_sig}", or "\code{NA}" specifying how each
    node along the dendrogram was handled by the iterative testing procedure}
\item{\code{ci_dat}}: {a matrix containing the cluster indices for the original
    data matrix passed to the constructor}
\item{\code{ci_sim}}: {a 3-dimensional array containing the simulated cluster
    indices at each subtree tested along the dendrogram}
\item{\code{p_emp}}: {a matrix containing the emprical p-values computed at each
    subtree tested along the dendrogram}
\item{\code{p_norm}}: {a matrix containing the Gaussian approximate p-values
    computed at each subtree tested along the dendrogram}
\item{\code{idx_hc}}: {a list of tuples containing the indices of clusters joined
    at each step of the hierarchical clustering procedure}
\item{\code{hc_dat}}: {a \code{hclust} object constructed from the original data matrix and
    arguments passed to the constructor}
}
}
\description{
Implements the Monte Carlo simulation based significance testing
procedure for hierarchical clustering described in Kimes et al. (Biometrics, 2017).
Statistical significance is evaluated at each node along the hierarchical
tree (dendrogram) starting from the root using a Gaussian null hypothesis test. 
A corresponding family-wise error rate (FWER) controlling procedure is
provided.
}
\details{
When possible, \code{Rclusterpp::Rclusterpp.hclust} should be used for
hierarchical clustering except in the case of clustering by Pearson
correlation (\code{dist="cor"}), for which we make use of
\code{WGCNA::cor} with the usual \code{stats::hclust}.

For standard minimum variance Ward's linkage clustering, if \code{rcpp} is
\code{TRUE}, specify "ward" for \code{linkage}. However, if \code{rcpp} is
\code{FALSE}, then "ward.D2" should be specified. See \code{stats::hclust}
for details on changes to the function since R >= 3.0.4. 

By default, p-values are computed for all nodes (\code{alpha = 1}).
If the procedure should instead terminate when no further nodes
meet a desired FWER control threshold, simply specify \code{alpha < 1}.
Controlling the FWER  may substantially speed up
the procedure by reducing the number of tests considered.

Even if the procedure is run using \code{alpha = 1}, the FWER cutoffs may still be
computed using \code{fwer_cutoff}. The FWER thresholding can also be visualized by
specifying \code{alpha} when calling \code{plot}.

The input \code{metric} can either be a character string specifying a metric
recognized by \code{dist()} or \code{"cor"} for Pearson correlation.
Alternatively, a user-defined dissimilarity function can be specified suing either
the \code{matmet} or \code{vecmet} parameter. If specified, \code{matmet} should be a
function which takes a matrix as input and returns an object of class \code{dist}
from the columns of the matrix. If specified, \code{vecmet} must be a function
which computes a real-valued dissimilarity value between two input vectors. This
function will be used to compute the dissimilarilty between columns of the data matrix.
If either \code{matmet} or \code{vecmet} is specified, \code{metric} will be ignored.
If both \code{matmet} and \code{vecmet} are specified, the function exit with an error.
Examples using \code{metric}, \code{matmet}, and \code{vecmet} are provided below.
}
\examples{
## using a string input to metric
data <- rbind(matrix(rnorm(100, mean = 2), ncol = 2),
              matrix(rnorm(100, mean = -1), ncol = 2))
shc_metric <- shc(data, metric = "cor", linkage = "average", alpha=0.05)
tail(shc_metric$p_norm, 10)

## using a function input to vecmet or any function not optimized for 
## computing dissimilarities for matrices will be incredibly slow
## should be avoided if possible
\dontrun{
vfun <- function(x, y) { 1 - cor(x, y) }
shc_vecmet <- shc(data, vecmet = vfun, linkage = "average", alpha=0.05)
tail(shc_vecmet$p_norm, 10)
}

## using a function input to matmet
mfun <- function(x) { as.dist(1-cor(x)) }
shc_mfun <- shc(data, matmet=mfun, linkage="average", alpha=0.05)
tail(shc_mfun$p_norm, 10)

}
\references{
\itemize{
    \item Kimes, P. K., Hayes, D. N., Liu Y., and Marron, J. S. (2016)
          Statistical significance for hierarchical clustering.
          pre-print available.
}
}
\seealso{
\code{\link{plot-shc}} \code{\link{diagnostic}}
}
\author{
Patrick Kimes
}
