% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{sigclust2-package}
\alias{sigclust2}
\alias{sigclust2-package}
\title{Statistical Significance for Clustering}
\description{
This package implements statistical significance of clustering procedures
based on the assumption that a cluster may be well approximated as a
single Gaussian distribution. The procedure is particularly well-suited for
the high-dimension low-sample size setting when the number of dimensions
greatly outnumbers the number of observations. The package builds substantially
on the \code{sigclust} package currently available through CRAN to include
additional options for null Gaussian estimation based on scaled PCA scores, and
more substantially, to test for significance in hierarchical clustering.
}
\author{
Patrick Kimes \email{patrick.kimes@gmail.com}
}
\keyword{package}

