% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary-shc}
\alias{summary-shc}
\alias{summary.shc}
\title{provide summary of shc object}
\usage{
\method{summary}{shc}(object, ...)
}
\arguments{
\item{object}{a \code{shc} object}

\item{...}{other parameters to be used by the function}
}
\description{
Provides information about \code{shc} object similar to call to \code{show}
}
\author{
Patrick Kimes
}

